Rename notification handler file
